A few fixes/simplifications to TrioManager
TrioManager was swallowing other task exceptions when a daemon task exited, raising
just a DaemonTaskExit. That is now fixed by raising a MultiError

Also gets rid of the coroutines responsible for marking the service as
stopping/cancelled as that is now done directly in the run() method.
That in turn allowed us to get rid of the system nursery altogether

Finally we no longer use an event to track cancellation and instead use
the task nursery's cancel_scope for that.