Fix task cancellation when the Service's run() method crashes
In some cases an exception raised on Service.run() can cause the manager
to hang forever waiting for all tasks to complete. Specifically, when
run() schedules a daemon task and then raises an exception itself, this
would happen:

 - _run_and_manage_task() catches the exception, then:
   - stores it in self._errors
   - triggers the service cancellation, and
   - removes the run() task from self._root_tasks
 - _handle_cancelled() kicks in
   - iterates over empty self._root_tasks, so doesn't cancel anything.
   - raise a LifecycleError because the daemon task is not done yet
 - _wait_all_tasks_done() never returns, so manager.run() doesn't either

This fixes that by changing _run_and_manage_task() to not remove
root tasks from self._root_tasks when they raise an exception and cause
the service to be cancelled.