Retry awaiting for stop() in background_asyncio_service()
In trinity, we inject a KeyboardInterrupt into a couroutine to tell it to stop,
and it will bubble up to background_asyncio_service(), whose finally: block
awaits for manager.stop(). However, for some of our services the manager.stop()
raises yet another KeyboardInterrupt, preventing trinity from shutting
down. I couldn't figure out where this 2nd KeyboardInterrupt comes from,
but it seems to happen always on the same services, and retrying that
always works (as in, the stop() call returns without error)